LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-none-expected.txt


FAIL Animation with animation-timeline:none holds current time at zero assert_equals: expected "100px" but got "101.640625px"
FAIL Animation with unknown timeline name holds current time at zero assert_equals: expected "100px" but got "101.640625px"
